pyproject.toml
[project]
name = "vqvae-vae-learn"
version = "0.1.0"
description = "VQ-VAE and VAE training with Hydra"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "numpy>=1.24",
    "Pillow>=9.0",
    "opencv-python>=4.0.0",
    "hydra-core>=1.3.0",
    "wandb>=0.15.0",
]
